versao inicial da calculadora
