+++ Principal.py
@@ -3,7 +3,7 @@ import numpy
 import tkinter as tk
 import customtkinter  as ctk 
 import string
-#                               Tamanho Da tela
+                                   # Tamanho Da tela
 numpy.set_printoptions(3, suppress=True)
 root = ctk.CTk()                  # Cria a nossa janela
 root._set_appearance_mode('dark')
